Update to using a ClickOneTwo app installer.
The Terminal-based installation still works, and uninstallation will remove both the older and the newer form of the utility.

Update the README accordingly. A link is still needed for the app installer on the app store, but that will have to wait for App Store approval.

+++ uninstall.sh
@@ -4,12 +4,28 @@ function fail {
 	exit 1
 }
 echo "Uninstalling launch agent..."
+# first remove the legacy location
 agent=~/Library/LaunchAgents/com.brotsky.backup.plist
+if [[ -f "$agent" ]]; then
+	launchctl unload "$agent"
+	rm -f "$agent"
+	[[ ! -f "$agent" ]] || fail "Failed to remove launch agent"
+fi
+# next remove the current location
+agent=~/Library/LaunchAgents/io.clickonetwo.easy-backup-and-eject.plist
 [[ -f "$agent" ]] && launchctl unload "$agent"
 rm -f "$agent"
 [[ ! -f "$agent" ]] || fail "Failed to remove launch agent"
 echo "Uninstalling script..."
+# first remove the legacy location
 scriptDir=~/.backup.brotsky.com
-rm -rf "$scriptDir"
-[[ ! -d "$scriptDir" ]] || fail "Failed to remove script directory"
-echo "com.brotsky.backup stopped and uninstalled"
+if [[ -d "$scriptDir" ]]; then
+	rm -rf "$scriptDir"
+	[[ ! -d "$scriptDir" ]] || fail "Failed to remove script directory"
+	echo "com.brotsky.backup stopped and uninstalled"
+fi
+# next remove the current script
+script=~/Library/"Application Support"/ClickOneTwo/easy-backup-and-eject.sh
+rm -rf "$script"
+[[ ! -d "$script" ]] || fail "Failed to remove script"
+echo "Easy Backup and Eject stopped and uninstalled"
